# apps/api/openapi.yaml
openapi: 3.0.0
info:
  title: CodeQual API
  version: 1.0.0
  description: AI-powered code review API for comprehensive pull request analysis
  termsOfService: https://codequal.com/terms
  contact:
    name: CodeQual Support
    email: support@codequal.com
    url: https://codequal.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.codequal.com/v1
    description: Production server
  - url: https://staging-api.codequal.com/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Development server
...
security:
  - ApiKeyAuth: []
paths:
...
  /analysis/history:
    get:
      summary: Get analysis history
      description: Retrieve pull request analysis history with filtering and pagination
      operationId: getAnalysisHistory
      tags:
        - Reports
      parameters:
        - name: limit
          in: query
          description: Number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: repositoryUrl
          in: query
          description: Filter by repository URL
          schema:
            type: string
            format: uri
      responses:
        '200':
          description: Analysis history retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisHistory'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication. Format ck_[32-character-hex]
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated users (internal use)
  schemas:
...
    AnalysisHistory:
      type: object
      properties:
        analyses:
          type: array
          items:
            type: object
            properties:
              analysisId:
                type: string
              repository:
                type: object
                properties:
                  url:
                    type: string
                  name:
                    type: string
              pr:
                type: object
                properties:
                  number:
                    type: integer
                  title:
                    type: string
                  author:
                    type: string
              timestamp:
                type: string
                format: date-time
              status:
                type: string
        pagination:
          type: object
          properties:
            total:
              type: integer
            offset:
              type: integer
            limit:
              type: integer
            hasMore:
              type: boolean
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code for programmatic handling
        details:
          type: object
          description: Additional error details
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: API key required
            code: AUTH_REQUIRED
...